requirements.txt
Bottleneck>=1.3.7
GPUtil>=1.4
hydra-core>=1.3.2
lightning>=2.0.0
neptune>=1.1.0
numpy>=1.24.2
protobuf==3.20.3
rich>=13.3.2
tensorboard>=2.12.0
torch>=2.0.0
torchmetrics>=0.11.4

# Not in package:
torchvision==0.15.1